`Removed NavigationOverlay component and related code from multiple files`
